+++ src/kimi_cli/llm.py
@@ -20,10 +20,11 @@ type ProviderType = Literal[
     "google_genai",  # for backward-compatibility, equals to `gemini`
     "gemini",
     "vertexai",
+    "_echo",
     "_chaos",
 ]
 
-type ModelCapability = Literal["image_in", "thinking"]
+type ModelCapability = Literal["image_in", "video_in", "thinking"]
 ALL_MODEL_CAPABILITIES: set[ModelCapability] = set(get_args(ModelCapability.__value__))
 
 
@@ -81,7 +82,10 @@ def create_llm(
     model: LLMModel,
     *,
     session_id: str | None = None,
-) -> LLM:
+) -> LLM | None:
+    if provider.type != "_echo" and (not provider.base_url or not model.model):
+        return None
+
     match provider.type:
         case "kimi":
             from kosong.chat_provider.kimi import Kimi
@@ -152,6 +156,10 @@ def create_llm(
                 api_key=provider.api_key.get_secret_value(),
                 vertexai=True,
             )
+        case "_echo":
+            from kosong.chat_provider.echo import EchoChatProvider
+
+            chat_provider = EchoChatProvider()
         case "_chaos":
             from kosong.chat_provider.chaos import ChaosChatProvider, ChaosConfig
             from kosong.chat_provider.kimi import Kimi